requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
scikit-image
rembg
onnxruntime-cpu
